requirements.txt
asgiref==3.10.0
certifi==2025.10.5
distlib==0.4.0
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
packaging==25.0
pillow==10.4.0
pipenv==2025.0.4
platformdirs==4.5.0
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.35.3
